requirements.txt
amqp==2.4.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.0
celery-with-redis==3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
colorama==0.4.1
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
gevent==1.4.0
greenlet==0.4.15
httpie==1.0.2
idna==2.8
kombu==4.5.0
lxml==4.3.3
psycopg2==2.8.1
pycparser==2.19
Pygments==2.3.1
python-dotenv==0.10.1
pytz==2019.1
redis==3.2.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
vine==1.3.0
